+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+<a name="1.0.5"></a>
+## [1.0.5](https://github.com/adonisjs/adonis-session/compare/v1.0.4...v1.0.5) (2017-07-31)
+
+
+### Bug Fixes
+
+* **cookie:** touch values cookie everytime ([06b02cc](https://github.com/adonisjs/adonis-session/commit/06b02cc))
+
+
+### Features
+
+* **command:** add config:session command ([89e49eb](https://github.com/adonisjs/adonis-session/commit/89e49eb))
+
+
+
 <a name="1.0.4"></a>
 ## [1.0.4](https://github.com/adonisjs/adonis-session/compare/v1.0.3...v1.0.4) (2017-07-22)